refactor(ui): dedicated folder
