+++ packages/twenty-ui/src/display/avatar-chip/components/AvatarChipLeftComponent.tsx
