+++ packages/twenty-ui/src/display/avatar-chip/types/AvatarChipsVariant.type.ts
