+++ packages/twenty-ui/src/display/index.ts
@@ -1,3 +1,9 @@
+export * from './avatar-chip/components/AvatarChip';
+export * from './avatar-chip/components/AvatarChipLeftComponent';
+export * from './avatar-chip/components/LinkAvatarChip';
+export * from './avatar-chip/types/AvatarChipsCommonProps.type';
+export * from './avatar-chip/types/AvatarChipsVariant.type';
+export * from './avatar-chip/utils/getAvatarChipsVariant.util';
 export * from './avatar/components/Avatar';
 export * from './avatar/components/AvatarGroup';
 export * from './avatar/components/states/isInvalidAvatarUrlState';
@@ -7,12 +13,6 @@ export * from './avatar/types/AvatarType';
 export * from './banner/components/Banner';
 export * from './checkmark/components/AnimatedCheckmark';
 export * from './checkmark/components/Checkmark';
-export * from './chip/avatar/components/AvatarChip';
-export * from './chip/avatar/components/AvatarChipLeftComponent';
-export * from './chip/avatar/components/LinkAvatarChip';
-export * from './chip/avatar/types/AvatarChipsCommonProps.type';
-export * from './chip/avatar/types/AvatarChipsVariant.type';
-export * from './chip/avatar/utils/getAvatarChipsVariant.util';
 export * from './chip/components/Chip';
 export * from './chip/components/LinkChip';
 export * from './color/components/ColorSample';
